+++ src/classes.js
@@ -1,4 +1,4 @@
-import { proxySocket } from "./upgraded-socket.js";
+import { createSocketProxy } from "./upgraded-socket.js";
 import { CLIENT, SERVER } from "./utils.js";
 
 const DEBUG = false;
@@ -73,7 +73,7 @@ export function formClientClass(ClientClass) {
 
     connectServerSocket(serverSocket) {
       if (DEBUG) console.log(`[ClientBase]  connected to server`);
-      this.server = proxySocket(CLIENT, SERVER, this, serverSocket);
+      this.server = createSocketProxy(CLIENT, SERVER, this, serverSocket);
       this.onConnect();
     }
 
@@ -116,7 +116,7 @@ export function formServerClass(ServerClass) {
     // the server.addClient(client) function for handling.
     async connectClientSocket(socket) {
       if (DEBUG) console.log(`[ServerBase] client connecting to server...`);
-      const client = proxySocket(SERVER, CLIENT, this, socket);
+      const client = createSocketProxy(SERVER, CLIENT, this, socket);
 
       // send the client its server id
       if (DEBUG) console.log(`[ServerBase] sending connection id`);